// packages/protoc-gen-grpc-js/src/gen/ts/service-edge-cases_grpc.ts
export const ServiceEmptyDefinition: Record<string, never> = {
};

/**
 * A service without any methods.
 *
 * @generated from service spec.ServiceEmpty
 */
export interface ServiceEmptyServer extends UntypedServiceImplementation {
}

/**
 * A service without any methods.
 *
 * @generated from service spec.ServiceEmpty
 */
export class ServiceEmptyClient extends Client {
}

/**
 * A service with method names that require escaping in generated code.
 *
 * @generated from service spec.ServiceNames
 */
export const ServiceNamesDefinition: {
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc spec.ServiceNames.Constructor
   */
  constructor$: MethodDefinition<Empty, Empty>;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc spec.ServiceNames.Delete
   */
  delete: MethodDefinition<Empty, Empty>;
} = {
  constructor$: {
    path: "/spec.ServiceNames/Constructor",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: Empty): Buffer => Buffer.from(toBinary(EmptySchema, value)),
    requestDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
    responseSerialize: (value: Empty): Buffer => Buffer.from(toBinary(EmptySchema, value)),
    responseDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
  },
  delete: {
    path: "/spec.ServiceNames/Delete",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: Empty): Buffer => Buffer.from(toBinary(EmptySchema, value)),
    requestDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
    responseSerialize: (value: Empty): Buffer => Buffer.from(toBinary(EmptySchema, value)),
    responseDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
  },
};
// ...
export class ServiceNamesClient extends Client {
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc spec.ServiceNames.Constructor
   */
  constructor$(request: Empty, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc spec.ServiceNames.Constructor
   */
  constructor$(request: Empty, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc spec.ServiceNames.Constructor
   */
  constructor$(request: Empty, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc spec.ServiceNames.Constructor
   */
  constructor$(request: Empty, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  constructor$(request: Empty, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Empty>, optionsOrCallback?: CallOptions | requestCallback<Empty>, callback?: requestCallback<Empty>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceNamesDefinition.constructor$.path,
      ServiceNamesDefinition.constructor$.requestSerialize,
      ServiceNamesDefinition.constructor$.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Empty>,
    );
  }
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc spec.ServiceNames.Delete
   */
  delete(request: Empty, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc spec.ServiceNames.Delete
   */
  delete(request: Empty, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc spec.ServiceNames.Delete
   */
  delete(request: Empty, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc spec.ServiceNames.Delete
   */
  delete(request: Empty, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  delete(request: Empty, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Empty>, optionsOrCallback?: CallOptions | requestCallback<Empty>, callback?: requestCallback<Empty>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceNamesDefinition.delete.path,
      ServiceNamesDefinition.delete.requestSerialize,
      ServiceNamesDefinition.delete.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Empty>,
    );
  }
}